--- app/components/server-navbar.tsx
@@ -1,12 +1,0 @@
-import { createPagesServerClient } from "@/lib/supabase";
-import Navbar from "./Navbar";
-
-export default async function ServerNavbar() {
-  const supabase = createPagesServerClient();
-  
-  const {
-    data: { user },
-  } = await supabase.auth.getUser();
-
-  return <Navbar initialUser={user} />;
-}
