+++ app/page.tsx
@@ -1,10 +1,15 @@
+"use client";
+
 import ClientPage from "@/app/client-page";
-import ServerNavbar from "@/components/server-navbar";
+import Navbar from "@/app/components/Navbar";
+import { useAuth } from "@/hooks/auth";
 
 export default function Home() {
+  const { user } = useAuth();
+  
   return (
     <>
-      <ServerNavbar />
+      <Navbar initialUser={user} />
       <ClientPage />
     </>
   );
