--- components/server-navbar.tsx
@@ -1,10 +1,0 @@
-import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
-import { cookies } from "next/headers";
-import Navbar from "@/app/components/Navbar";
-
-export default async function ServerNavbar() {
-  const supabase = createServerComponentClient({ cookies });
-  const { data: { user } } = await supabase.auth.getUser();
-  
-  return <Navbar initialUser={user} />;
-}
