+++ lib/supabase.ts
@@ -1,6 +1,5 @@
 // Client-side exports
 export { createClient } from './supabase/client'
-export { createPagesServerClient } from './supabase/pages'
 
 // Server-side only exports
 export { createAdminClient } from './supabase/admin'
