--- lib/supabase/index.ts
@@ -1,9 +1,0 @@
-import { createClient as createSupabaseClient } from '@supabase/supabase-js'
-import type { SupabaseClient } from '@supabase/supabase-js'
-
-const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL!
-const supabaseAnonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
-
-export const createClient = (): SupabaseClient => {
-  return createSupabaseClient(supabaseUrl, supabaseAnonKey)
-}
